+++ tests/test_trans_irrc.rs
@@ -111,6 +111,9 @@ fn test_irrc_trans_lpf_loop() {
 fn build( freq: f64 ) -> Circuit {
     let mut ckt = Circuit::new();
 
+    ckt.add_node("1");
+    ckt.add_node("2");
+
     // 10V Voltage Source
     ckt.elements.push(
         Element::Isin(CurrentSourceSine{p: 0, n: 1, vo: 0.0, va: 10.0, freq: freq}),
@@ -126,6 +129,8 @@ fn build( freq: f64 ) -> Circuit {
     ckt.elements.push(
         Element::C(Capacitor{a: 2, b: 0, value: 0.032e-6}),
     );
+
+    ckt.build_node_id_lut();
     ckt
 }