+++ tests/test_trans_irrc_hpf.rs
@@ -78,6 +78,9 @@ fn test_irrc_trans_hpf_10kHz() {
 fn build( freq: f64 ) -> Circuit {
     let mut ckt = Circuit::new();
 
+    ckt.add_node("1");
+    ckt.add_node("2");
+
     // 10V Voltage Source
     ckt.elements.push(
         Element::Isin(CurrentSourceSine{p: 0, n: 1, vo: 0.0, va: 10.0, freq: freq}),
@@ -93,6 +96,8 @@ fn build( freq: f64 ) -> Circuit {
     ckt.elements.push(
         Element::R(Resistor{a: 2, b: 0, value: 1.0e3}),
     );
+
+    ckt.build_node_id_lut();
     ckt
 }